// src/types/language.ts
export type LanguagesCode = "en" | "bn" | "tr" | "cn" | "kr";

// Language type
export interface Language {
  title: string;
  code: LanguagesCode;
}